refact: move Event Emitter to /util
